requirements.txt
Flask
pytest
coverage
pyserial